Backport 824a5e4c605d4aee55252bce5364fa01de525e1b
